fix: Act on four of five review comments
isdigit accepted what int() then rejected. A superscript passes isdigit, so a
target could clear the boundary check and raise ValueError inside the drainer -
after _drain_once had already removed the pair, losing the shoutout. isascii and
isdecimal is the pair that matches: isdecimal refuses the superscript, isascii
refuses an Arabic-Indic numeral that int() would otherwise take happily.

resume() said it did not raise and then converted the broadcaster id outside its
own guard, so a missing or non-numeric setting raised past the report and into a
gather(return_exceptions=True) that swallows it - startup with no queue and
nothing said. The conversion moved inside, and the guard widened to the two
errors it can actually produce.

The hourly re-check started with an empty known-broken map, so its first pass
reported everything the startup check had already reported an hour earlier. The
fix was to delete the startup check: an interval loop runs its first pass
immediately, so it always was the startup check as well. The shared wording went
with it, into the only module still using it, which also drops a cog importing
from init.

_deliver counted a channel it could not resolve but not a send that raised on
one it could. Counting before the attempt covers both, and the count still
clears only when something has actually arrived.

Not changed: the fifth comment says downgrading 0004 would delete a
command_failed row somebody had added by hand before the migration ran. Upgrade
tolerates a conflict because db/README requires it; downgrade removing what
upgrade adds is what 0003 does too, and no migration here guards against a
hand-added row of the same key. Guarding this one would be machinery for a case
the repo does not have.

Twitch-relayed login and id are now logged as %r, like every other place that
relays external text.

Verified: five id shapes queue or refuse to match what int() plus ascii accepts;
a bad broadcaster id reports rather than raising; the startup check is gone and
the loop is an interval loop; three sends that raise are counted and reported on
recovery.
