+++ init/bot_init.py
@@ -24,44 +24,10 @@ async def resume_stream_session() -> None:
     await stream_session.resume()
 
 
-def undeliverable_summary(broken: dict[str, str]) -> str:
-    """The admin-channel wording, shared by the startup check and the loop."""
-    detail = "\n".join(f"- {identity}: {reason}" for identity, reason in broken.items())
-    return (
-        f"{len(broken)} Twitch subscription(s) will not deliver:\n{detail}\n"
-        "/subscribe replaces the stream.online and stream.offline pair; the"
-        " rest are registered outside the bot."
-    )
-
-
-async def check_subscriptions() -> None:
-    """Say at startup when a Twitch subscription will not deliver.
-
-    The only way Twitch reports a subscription it has disabled is by calling the
-    webhook, which is the very thing that is not working, so nothing else in the
-    bot would ever find out. A stream alert that silently stopped existing is
-    the most expensive failure here and the quietest.
-    """
-    from errors import notify, report
-    from services.twitch.api import broken_subscriptions
-    from services.twitch.helix import HelixError
-
-    try:
-        broken = await broken_subscriptions()
-    except HelixError as e:
-        # gather(return_exceptions=True) below would swallow this silently.
-        await report(e, "Could not check the Twitch subscriptions at startup")
-        return
-
-    if broken:
-        await notify(undeliverable_summary(broken))
-
-
 async def run_background_tasks() -> None:
     await asyncio.gather(
         restart_live_alert_tasks(),
         resume_stream_session(),
-        check_subscriptions(),
         return_exceptions=True,
     )
